count and status message fixes
if count is not given in the conf for the required files then the file
has to be present at least once. if a count is given the file has to be
found exactly so many times.

fixed printing of full status, where only the last status messages are
printed.